Spreadsheet: use forced-interruptible i/o for file parsing
this involves creating a fileobj wrapper that will forcibly
inject interruption points when being read from, and using
that wrapper for the uploaded fileobj directly *and*
monkeypatching openpyxl to use our ZipFile subclass that
wraps opened extracted fileobjs in these interruptible
wrappers.

this is necessary because the fileobjs we're dealing with
are in-memory and don't perform any actual i/o, which is
where eventlet would normally be able to interrupt their
processing.