Unpin dependencies in `requirements.in`
We used to freeze our requirements with a custom script. This script
needed version numbers in the `.in` file (then called
`requirements-app.txt`) so it could check that the `requirements.txt`
file was up to date.

Here’s where it was doing that:
https://github.com/alphagov/notifications-api/pull/1938/files#diff-76ed074a9305c04054cdebb9e9aad2d818052b07091de1f20cad0bbac34ffb52R107-R112

Since then we’ve moved to `pip-compile` and the `uv pip-compile` to do
the same thing.

These are much better at not modifying the `requirements.txt` file
unless they have to.

We also moved to checking that the requirements files is correct by
re-generating it, rather than comparing it:
https://github.com/alphagov/notifications-aws/pull/1207/files

Doing this will make it easier for us (and Dependabot) to bump the
versions of things because we’ll only need to do so in one place.