Update UpdateContent tests
This updates tests with the new module.
What tests test has not been changed.

The change is the way tests are executed and
time in them advanced.

Most of our test have an immediate response
so for them we have mock response and
mock of the fetch implementation.

For test where we test a response that takes
longer we mock a delay response in
`mockResponseDelay` and override the
default implementation.

To working with the Promise nature of fetch, we
advance time with `jest.advanceTimersByTimeAsync`
https://jestjs.io/docs/jest-object#jestadvancetimersbytimeasyncmstorun
that allows any scheduled promise callbacks
to execute before running the timers.
Note: `advanceTimersByTimeAsync` is available as
of 29.5 as of Jest and 'modern' fake timers have
been the default since 27.0.