Add check for wrong type of dashes in content
Similar to how we check for the wrong type of quotes.

Dashes to separate clauses – like these – should always be en dashes.

Dashes should not be used to separate ranges, in other words it should
be ‘2025 to 2026’ not ‘2025–2026’ or ‘2025-2026’.

Em dashes should not be used at all (they are a little anachronistic).