Update `confirmed_email_sender_name` attribute
If a user clicks "Save" on the "Choose a ‘from’ name" page we now update
the value of `confirmed_email_sender_name` to True regardless of whether
they have chosen to use the name of their service or have entered a
custom name.

Confirming the email from name will become a compulsory part of going live
for services planning on sending emails. It will also become necessary
for live services who have emails turn off and who want to then turn it
on.

This is the only point at which we change the value of
`confirmed_email_sender_name` - if a service's name then changes we
don't set this value to False. This can be revisited but the decision to
keep it set as True was made because we think that services will have
already got the idea that changing the service name changes the email
sender name and it could be a little frustrating to make them
re-confirm, especially if the service is live.