Only add placeholder if it doesn’t exist
If the template already has a placehodler for the link (because the user
for some reason thought they had to create it first) then it’s better
we don’t put a second link in their template.

In making the test for this case more robust I found some wrong data
being passed around which I had to fix.