add email files as attribute to deserialized EmailPreviewTemplate object
the view_template view uses an instance of `EmailPreviewTemplate` to render a preview of a template. This adds a list of SerializedTemplateEmailFile dicts `https://github.com/alphagov/notifications-api/blob/main/app/models.py#L1304-L1310` as an attribute so that this view can render counts, links to preview files etc.