Add default value for template email files bucket in dev
This matches what we do for other S3 buckets and means that you don't
need to manually set the environment variable when running locally in
order for the feature to work.

It also adds the `S3_BUCKET_DOCUMENT_DOWNLOAD_LONG_TERM_FILE_STORAGE
enviroment variable to the test config for consistency with the other S3
environment variables.